backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
minio==7.2.0
python-docx==1.1.0
weasyprint==61.0
pillow==10.2.0
pandas==2.2.0
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
celery==5.3.6
playwright>=1.40.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.0.0
pytest-asyncio==0.23.0
